db/transport: widen DB put offset/size to uint64 (symmetry with get path)
Latent >4GB correctness for the DB write-back/put path; does not resolve the
conv-2d megalarge corruption (which is bit-identical across all transport
byte-width fixes, so it is NOT a transport overflow). gemm/2mm unaffected.

+++ libs/include/internal/arts/memory/db.h
@@ -87,8 +87,8 @@ bool arts_add_db_duplicate_ex(struct arts_db_s *db, unsigned int rank,
                               uint32_t flags, bool *on_head);
 void prep_dbs(unsigned int depc, arts_edt_dep_t *depv, bool gpu);
 void internal_put_in_db(void *ptr, arts_guid_t edt_guid, arts_guid_t db_guid,
-                        unsigned int slot, unsigned int offset,
-                        unsigned int size, arts_guid_t epoch_guid,
+                        unsigned int slot, uint64_t offset,
+                        uint64_t size, arts_guid_t epoch_guid,
                         unsigned int rank);
 void arts_get_from_db_at_ex(arts_guid_t edt_guid, arts_guid_t db_guid,
                             unsigned int slot, uint64_t offset,
